Added machines files to recognize tokens.
This file exports an array of function pointers called "machines".  Each one
of these takes a pointer to a string, and returns a MachineResult.

This currently is just the updated pointer, but can be changed to anything
desired (such as the element of the symbol table, the number itself, as well
as the token type and any attributes.)