Added translation table to strings.
Oh my. There has got to be a better way to do this. I hate non-DRY code,
and this screams of it. What happens when somebody adds a new constant,
expecting it to just work? Maybe I could try including src/types.h in
the redistributed package, and parsing it, using those instead.