Initialization of variables
Really odd. For some reason sfSymbolTable, if uninitialized, was always
pointing to a spot in memory containing the string "2'". This meant that
every time I ran the program it would open that file and output to it.
This is the kind of bug that makes me miss python. Why didn't -Wall
catch this?